requirements.txt
Django==1.4.1
Fabric==1.4.3
boto==2.6.0
django-storages==1.1.5
django-tastypie==0.9.11
django-tastypie-mongoengine==0.3
drest==0.9.10
elementtree==1.2.7-20070827-preview
gunicorn==0.14.6
httplib2==0.7.6
mimeparse==0.1.3
mongoengine==0.7.5
pycrypto==2.6
pymongo==2.3
python-dateutil==1.5
requests==0.14.1
slumber==0.4.2
ssh==1.7.14
wsgiref==0.1.2
